Add files via upload
Fiddled around with optimizer and learning rate.
The best model performance on the unseen test images(with minimal overfitting) was with ADAM optimizer (default learning rate of 0.001). SGD optimizer ,overall, was not able to perform as well as the ADAM optimizer and could not generalize well over unseen X ray data given the training epochs. SGD tends to fluctuate between local minimas due to randomness in its descent. So it may be trained for a period >30 epochs to possibly get better results .
Logistic regression is a simpler model compared to CNN, but its results were comparable to it with very few layers